+++ src/components/user/AddUser.jsx
@@ -3,23 +3,15 @@ import { useEffect, useState } from "react";
 
 import ExpandLessIcon from "@mui/icons-material/ExpandLess";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import Visibility from "@mui/icons-material/Visibility";
-import VisibilityOff from "@mui/icons-material/VisibilityOff";
-import Button from "@mui/material/Button";
 import Card from "@mui/material/Card";
 import CardContent from "@mui/material/CardContent";
 import CardHeader from "@mui/material/CardHeader";
-import Checkbox from "@mui/material/Checkbox";
-import FormControlLabel from "@mui/material/FormControlLabel";
-import FormGroup from "@mui/material/FormGroup";
-import Grid from "@mui/material/Grid";
 import IconButton from "@mui/material/IconButton";
-import InputAdornment from "@mui/material/InputAdornment";
-import TextField from "@mui/material/TextField";
 import dao from "../../ajax/dao";
 import ValidateAddUser from "../../validation/ValidateAddUser";
 import AlertBox from "../common/AlertBox";
 import ConfirmationDialog from "../common/ConfirmationDialog";
+import AddUserForm from "./AddUserForm";
 
 export default function AddUser({ getAllUsers }) {
   const [isCardExpanded, setIsCardExpanded] = useState(false);
@@ -43,17 +35,6 @@ export default function AddUser({ getAllUsers }) {
     isPlanner: 0,
     isStatist: 0,
   });
-
-  const [showPassword, setShowPassword] = useState(false);
-
-  const handlePasswordVisibility = () => {
-    setShowPassword(!showPassword);
-  };
-
-  const handleCheckUserRole = (rolename) => (event) => {
-    const updatedValue = event.target.checked ? 1 : 0;
-    formik.setFieldValue(rolename, updatedValue);
-  };
 
   useEffect(() => {
     getAllUsers();
@@ -136,111 +117,11 @@ export default function AddUser({ getAllUsers }) {
             }
           />
           {isCardExpanded && (
-            <form onSubmit={formik.handleSubmit}>
-              <Grid container spacing={2}>
-                <Grid item xs={12}>
-                  <TextField
-                    className="formTextInput"
-                    error={
-                      formik.touched.email && formik.errors.email ? true : false
-                    }
-                    name="email"
-                    placeholder="Email..."
-                    label="Email"
-                    value={formik.values.email}
-                    onChange={formik.handleChange("email")}
-                    onBlur={formik.handleBlur("email")}
-                    helperText={
-                      formik.touched.email && formik.errors.email
-                        ? formik.errors.email
-                        : null
-                    }
-                  />
-                </Grid>
-                <Grid item xs={12}>
-                  <TextField
-                    name="password"
-                    placeholder="Password..."
-                    label="Password"
-                    className="formTextInput"
-                    error={
-                      formik.touched.password && !formik.values.password
-                        ? true
-                        : false
-                    }
-                    value={formik.values.password}
-                    onChange={formik.handleChange("password")}
-                    type={showPassword ? "text" : "password"}
-                    helperText={
-                      formik.touched.password && !formik.values.password
-                        ? "Password is a required field"
-                        : null
-                    }
-                    InputProps={{
-                      endAdornment: (
-                        <InputAdornment position="end">
-                          <IconButton
-                            onClick={handlePasswordVisibility}
-                            edge="end"
-                            style={{ backgroundColor: "transparent" }}
-                          >
-                            {showPassword ? <Visibility /> : <VisibilityOff />}
-                          </IconButton>
-                        </InputAdornment>
-                      ),
-                    }}
-                  />
-                </Grid>
-                <Grid item xs={12} sm={6} md={4}>
-                  <div className="errorMessage">
-                    {(formik.touched.isAdmin ||
-                      formik.touched.isPlanner ||
-                      formik.touched.isStatist) &&
-                    Boolean(formik.errors.roles) === true
-                      ? formik.errors.roles
-                      : null}
-                  </div>
-                  <FormGroup>
-                    <FormControlLabel
-                      control={<Checkbox />}
-                      label="Admin"
-                      labelPlacement="start"
-                      className="formCheckBoxButtons"
-                      name="isAdmin"
-                      checked={formik.values.isAdmin === 1}
-                      onChange={handleCheckUserRole("isAdmin")}
-                    />
-                    <FormControlLabel
-                      control={<Checkbox />}
-                      label="Planner"
-                      labelPlacement="start"
-                      className="formCheckBoxButtons"
-                      name="isPlanner"
-                      checked={formik.values.isPlanner === 1}
-                      onChange={handleCheckUserRole("isPlanner")}
-                    />
-                    <FormControlLabel
-                      control={<Checkbox />}
-                      label="Statist"
-                      labelPlacement="start"
-                      className="formCheckBoxButtons"
-                      name="isStatist"
-                      checked={formik.values.isStatist === 1}
-                      onChange={handleCheckUserRole("isStatist")}
-                    />
-                  </FormGroup>
-                  <Button
-                    type="submit"
-                    variant="addComponentFormButton"
-                    onClick={() => {
-                      setInitialUser(formik.values);
-                    }}
-                  >
-                    Add User
-                  </Button>
-                </Grid>
-              </Grid>
-            </form>
+            <AddUserForm
+              formik={formik}
+              submitValues={formik.values}
+              setInitialUser={setInitialUser}
+            />
           )}
         </CardContent>
       </Card>
